Move vec2 to independant file
